# internal/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.TaskAttemptAnswerRequest:
    properties:
      selected_option_ids:
        items:
          type: string
        type: array
      task_id:
        type: string
      text_answer:
        type: string
    type: object
  dto.TaskAttemptAnswerResponse:
    properties:
      is_correct:
        type: boolean
      selected_option_ids:
        items:
          type: string
        type: array
      task_id:
        type: string
      text_answer:
        type: string
    type: object
  dto.TaskAttemptListResponse:
    properties:
      task_attempts:
        items:
          $ref: '#/definitions/dto.TaskAttemptResponse'
        type: array
    type: object
  dto.TaskAttemptRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.TaskAttemptAnswerRequest'
        type: array
      course_id:
        type: string
      module_id:
        type: string
      user_id:
        type: string
    type: object
  dto.TaskAttemptResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.TaskAttemptAnswerResponse'
        type: array
      attempt_number:
        type: integer
      correct_answers:
        type: integer
      course_id:
        type: string
      id:
        type: string
      module_id:
        type: string
      score:
        type: number
      total_questions:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API для пользователей Course Service
  title: Course Service API
  version: "1.0"
paths:
...
  /task-attempts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ответы пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TaskAttemptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отправить попытку прохождения теста
      tags:
      - Task Attempts
  /task-attempts/{id}:
    get:
      parameters:
      - description: Attempt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить попытку
      tags:
      - Task Attempts
  /task-attempts/courses/{course_id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить все попытки по курсу
      tags:
      - Task Attempts
  /task-attempts/modules/{module_id}:
    get:
      parameters:
      - description: Module ID
        in: path
        name: module_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить все попытки по модулю
      tags:
      - Task Attempts
  /task-attempts/my/courses/{course_id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить мои попытки по курсу
      tags:
      - Task Attempts
  /task-attempts/my/modules/{module_id}:
    get:
      parameters:
      - description: Module ID
        in: path
        name: module_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить мои попытки по модулю
      tags:
      - Task Attempts
  /task-attempts/users/{user_id}/courses/{course_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить попытки пользователя по курсу
      tags:
      - Task Attempts
  /task-attempts/users/{user_id}/modules/{module_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Module ID
        in: path
        name: module_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskAttemptListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить попытки пользователя по модулю
      tags:
      - Task Attempts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"